requirements.txt
streamlit
google-generativeai
pandas
python-dotenv
openpyxlstreamlit
google-generativeai
pandas
python-dotenv
openpyxl